Typecast timeout parameter.
For some reason VxWorks 6.9 headers define WAIT_FOREVER as (-1U) instead of
(-1), like every previous version. This generates signed/unsigned
comparison warnings. This kludge should still work, but it's unfortunate
that it's necessary.

+++ util.cpp
@@ -7,8 +7,8 @@ using namespace vwpp::v2_7;
 
 int ms_to_tick(int const v)
 {
-    if (v == WAIT_FOREVER)
-	return WAIT_FOREVER;
+    if (v == (int) WAIT_FOREVER)
+	return (int) WAIT_FOREVER;
     else
 	return (std::max(v, 0) * ::sysClkRateGet() + 999) / 1000;
 }
